+++ lib/visual_snapshots/run-android-dextrose.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-npx appium &
-APPIUM_PID=$!
 LOCAL=true npx fructose-tunnel
 emulator @Nexus_5X_API_25 -no-boot-anim &
 EMU_PID=$!
@@ -14,8 +12,7 @@ npx lerna run dextrose-stories --since
 npx compile-tests -d fructose
 npx rnstl --searchDir ./packages --pattern './*/*.dextrose.tmp.js' --outputFile ./fructose/components.js
 npx react-native start --root fructose --projectRoots $(pwd)/fructose,$(pwd) &
-react-native run-android --variant=debug --no-packager
-LOGLEVEL=verbose npx run dextrose --config ./dextrose/dextrose.android.js --timeout 600000  --snapshotWait 2000
+npx react-native run-android --variant=release --no-packager
+npx dextrose run --config ./dextrose/dextrose.android.js --timeout 600000  --snapshotWait 2000
 npx lerna run dextrose-clean
 kill -9 $EMU_PID
-kill -9 $APPIUM_PID