+++ lib/visual_snapshots/run-ios-dextrose.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-appium &
-APPIUM_PID=$!
 xcrun simctl boot 'iPhone 7'
 npx lerna run dextrose-clean
 npx lerna run dextrose-stories --since
@@ -12,8 +10,7 @@ export FRC=$CWD/fructose
 npx react-native start --root fructose --projectRoots $(pwd)/fructose,$(pwd) &
 react-native run-ios --no-packager
 PACKAGER_PID=$!
-LOGLEVEL=verbose npx run dextrose --config ./dextrose/dextrose.ios.js --snapshotWait 2000
+LOGLEVEL=verbose npx dextrose run --config ./dextrose/dextrose.ios.js --snapshotWait 2000
 npx lerna run dextrose-clean
 kill -9 $PACKAGER_PID
 xcrun simctl shutdown booted
-kill -9 $APPIUM_PID