+++ packages/article-flag/__tests__/ios/article-flag.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article Flag test on ios", () => {
+  shared();
+});