+++ packages/article-summary/article-summary.js
@@ -1,8 +1,12 @@
 import React from "react";
-import { Text, View, Platform } from "react-native";
+import { Text, View } from "react-native";
 import PropTypes from "prop-types";
+
 import { renderTrees, treePropType } from "@times-components/markup";
 import DatePublication from "@times-components/date-publication";
+import ArticleByline from "@times-components/article-byline";
+
+import ArticleSummaryHeadline from "./article-summary-headline";
 import renderer from "./article-summary-renderer";
 
 const styles = {
@@ -14,17 +18,6 @@ const styles = {
     marginBottom: 2,
     letterSpacing: 1
   },
-  headline: {
-    color: "#333333",
-    fontSize: 22,
-    lineHeight: 22,
-    marginBottom: 6,
-    fontFamily: "TimesModern-Bold",
-    fontWeight: "400",
-    ...Platform.select({
-      web: { WebkitFontSmoothing: "auto" }
-    })
-  },
   text: {
     color: "#696969",
     fontSize: 14,
@@ -33,12 +26,12 @@ const styles = {
     marginBottom: 10,
     flexWrap: "wrap"
   },
-  datePublication: {
+  metaText: {
     color: "#696969",
     fontSize: 13,
     lineHeight: 15,
     fontFamily: "GillSansMTStd-Medium",
-    marginBottom: 10
+    marginBottom: 5
   }
 };
 
@@ -58,21 +51,45 @@ const summarise = text => {
 };
 
 const ArticleSummary = props => {
-  const { label, headline, text, date, publication } = props;
+  const {
+    label,
+    headline,
+    hasResponsiveHeadline,
+    text,
+    date,
+    publication,
+    showPublication,
+    byline
+  } = props;
   const summary = summarise(text);
   const labelText = label && label.toUpperCase && label.toUpperCase();
+
   return (
     <View style={styles.container}>
-      <Text style={styles.label}>{labelText}</Text>
-      <Text style={styles.headline}>{headline}</Text>
+      {label ? <Text style={styles.label}>{labelText}</Text> : null}
+      {headline ? (
+        <ArticleSummaryHeadline
+          hasResponsiveHeadline={hasResponsiveHeadline}
+          headline={headline}
+        />
+      ) : null}
       <Text style={styles.text}>{renderTrees(summary, renderer)}</Text>
       <Text
-        style={styles.datePublication}
+        style={styles.metaText}
         accessibilityLabel="datePublication"
         testID="datePublication"
       >
-        <DatePublication date={date} publication={publication} />
+        <DatePublication
+          date={date}
+          publication={publication}
+          showPublication={showPublication}
+        />
       </Text>
+      {byline.length ? (
+        <Text style={styles.metaText}>
+          <ArticleByline ast={byline} />
+        </Text>
+      ) : null}
     </View>
   );
 };
@@ -80,17 +97,23 @@ const ArticleSummary = props => {
 ArticleSummary.propTypes = {
   label: PropTypes.string,
   headline: PropTypes.string,
+  hasResponsiveHeadline: PropTypes.bool,
   text: PropTypes.arrayOf(treePropType),
   date: DatePublication.propTypes.date,
-  publication: DatePublication.propTypes.publication
+  publication: DatePublication.propTypes.publication,
+  showPublication: DatePublication.propTypes.showPublication,
+  byline: PropTypes.arrayOf(treePropType)
 };
 
 ArticleSummary.defaultProps = {
   label: "",
   headline: "",
+  hasResponsiveHeadline: false,
   text: [],
   date: null,
-  publication: DatePublication.defaultProps.publication
+  publication: DatePublication.defaultProps.publication,
+  showPublication: DatePublication.defaultProps.showPublication,
+  byline: []
 };
 
 export default ArticleSummary;