+++ packages/article-summary/article-summary.stories.js
@@ -2,19 +2,22 @@ import React from "react";
 import { View } from "react-native";
 import { storiesOf } from "dextrose/storiesOfOverloader";
 import ArticleSummary from "./article-summary";
+
+import defaultFixture from "./fixtures/default.json";
+import articleMultiFixture from "./fixtures/article-multi.json";
+import noBylineFixture from "./fixtures/no-byline.json";
+import noLabelFixture from "./fixtures/no-label.json";
 import reviewFixture from "./fixtures/review.json";
-import articleMulti from "./fixtures/article-multi.json";
-import articleSingle from "./fixtures/article-single.json";
 
 const story = m => <View style={{ padding: 20 }}>{m}</View>;
 
 storiesOf("ArticleSummary", module)
-  .add("Paragraph Summary (multiple)", () =>
-    story(<ArticleSummary {...articleMulti} />)
+  .add("Default", () => story(<ArticleSummary {...defaultFixture} />))
+  .add("Summary with multiple paragraphs", () =>
+    story(<ArticleSummary {...articleMultiFixture} />)
   )
-  .add("Paragraph Summary (single)", () =>
-    story(<ArticleSummary {...articleSingle} />)
-  )
+  .add("No byline", () => story(<ArticleSummary {...noBylineFixture} />))
+  .add("No label", () => story(<ArticleSummary {...noLabelFixture} />))
   .add("Review/Rating Summary", () =>
     story(<ArticleSummary {...reviewFixture} />)
   );