+++ packages/caption/__tests__/web/caption.web.test.js
@@ -1,0 +1,6 @@
+import "react-native";
+import shared from "../shared";
+
+describe("Caption on Web", () => {
+  shared();
+});