+++ packages/slice/__tests__/web/slice.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Slice test on web", () => {
+  shared();
+});