+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.11.17"></a>
+## [0.11.17](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.16...@times-components/article-image@0.11.17) (2018-01-30)
+
+
+### Bug Fixes
+
+* android dextrose ([#569](https://github.com/newsuk/times-components/issues/569)) ([b5b74fa](https://github.com/newsuk/times-components/commit/b5b74fa))
+
+
+
+
+<a name="0.11.16"></a>
+## [0.11.16](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.15...@times-components/article-image@0.11.16) (2018-01-30)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
+<a name="0.11.15"></a>
+## [0.11.15](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.14...@times-components/article-image@0.11.15) (2018-01-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
+<a name="0.11.14"></a>
+## [0.11.14](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.13...@times-components/article-image@0.11.14) (2018-01-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
+<a name="0.11.13"></a>
+## [0.11.13](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.12...@times-components/article-image@0.11.13) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
+<a name="0.11.12"></a>
+## [0.11.12](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.11...@times-components/article-image@0.11.12) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
+<a name="0.11.11"></a>
+## [0.11.11](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.10...@times-components/article-image@0.11.11) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-image
+
 <a name="0.11.10"></a>
 ## [0.11.10](https://github.com/newsuk/times-components/compare/@times-components/article-image@0.11.9...@times-components/article-image@0.11.10) (2018-01-19)