+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.17"></a>
+## [0.7.17](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.16...@times-components/error-view@0.7.17) (2018-01-30)
+
+
+### Bug Fixes
+
+* android dextrose ([#569](https://github.com/newsuk/times-components/issues/569)) ([b5b74fa](https://github.com/newsuk/times-components/commit/b5b74fa))
+
+
+
+
+<a name="0.7.16"></a>
+## [0.7.16](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.15...@times-components/error-view@0.7.16) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.7.15"></a>
 ## [0.7.15](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.7.14...@times-components/error-view@0.7.15) (2018-01-19)