+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.16.6"></a>
+## [0.16.6](https://github.com/newsuk/times-components/compare/@times-components/markup@0.16.5...@times-components/markup@0.16.6) (2018-01-30)
+
+
+### Bug Fixes
+
+* android dextrose ([#569](https://github.com/newsuk/times-components/issues/569)) ([b5b74fa](https://github.com/newsuk/times-components/commit/b5b74fa))
+
+
+
+
+<a name="0.16.5"></a>
+## [0.16.5](https://github.com/newsuk/times-components/compare/@times-components/markup@0.16.4...@times-components/markup@0.16.5) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
+<a name="0.16.4"></a>
+## [0.16.4](https://github.com/newsuk/times-components/compare/@times-components/markup@0.16.3...@times-components/markup@0.16.4) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/markup
+
 <a name="0.16.3"></a>
 ## [0.16.3](https://github.com/newsuk/times-components/compare/@times-components/markup@0.16.2...@times-components/markup@0.16.3) (2018-01-19)