+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.37"></a>
+## [0.12.37](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.36...@times-components/pagination@0.12.37) (2018-01-30)
+
+
+### Bug Fixes
+
+* android dextrose ([#569](https://github.com/newsuk/times-components/issues/569)) ([b5b74fa](https://github.com/newsuk/times-components/commit/b5b74fa))
+
+
+
+
+<a name="0.12.36"></a>
+## [0.12.36](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.35...@times-components/pagination@0.12.36) (2018-01-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.35"></a>
+## [0.12.35](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.34...@times-components/pagination@0.12.35) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
+<a name="0.12.34"></a>
+## [0.12.34](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.33...@times-components/pagination@0.12.34) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/pagination
+
 <a name="0.12.33"></a>
 ## [0.12.33](https://github.com/newsuk/times-components/compare/@times-components/pagination@0.12.31...@times-components/pagination@0.12.33) (2018-01-24)