+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/utils@0.1.7...@times-components/utils@0.2.0) (2018-01-30)
+
+
+### Features
+
+* Update to Articles Provider to Include Related Articles ([910fc2d](https://github.com/newsuk/times-components/commit/910fc2d))
+
+
+
+
+<a name="0.1.7"></a>
+## [0.1.7](https://github.com/newsuk/times-components/compare/@times-components/utils@0.1.6...@times-components/utils@0.1.7) (2018-01-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/utils
+
+<a name="0.1.6"></a>
+## [0.1.6](https://github.com/newsuk/times-components/compare/@times-components/utils@0.1.5...@times-components/utils@0.1.6) (2018-01-26)
+
+
+
+
+**Note:** Version bump only for package @times-components/utils
+
 <a name="0.1.5"></a>
 ## [0.1.5](https://github.com/newsuk/times-components/compare/@times-components/utils@0.1.4...@times-components/utils@0.1.5) (2018-01-19)